requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

psutil>=5.9.0
paho-mqtt>=1.6.1
python-dateutil>=2.8.2
